requirements.txt
ccxt
yfinance
pandas
# pandas_ta # Incompatible with Python 3.14
python-dotenv
python-dotenv
requests
pycoingecko
pyTelegramBotAPI
python-binance
alpaca-py
PySocks
pycares>=4.0.0
aiodns<3.2.0
openai
